+++ src/ImportDefinitionsBundle/DependencyInjection/Compiler/ExportProviderRegistryCompilerPass.php
@@ -1,0 +1,30 @@
+<?php
+/**
+ * Import Definitions.
+ *
+ * LICENSE
+ *
+ * This source file is subject to the GNU General Public License version 3 (GPLv3)
+ * For the full copyright and license information, please view the LICENSE.md and gpl-3.0.txt
+ * files that are distributed with this source code.
+ *
+ * @copyright  Copyright (c) 2016-2018 w-vision AG (https://www.w-vision.ch)
+ * @license    https://github.com/w-vision/ImportDefinitions/blob/master/gpl-3.0.txt GNU General Public License version 3 (GPLv3)
+ */
+
+namespace ImportDefinitionsBundle\DependencyInjection\Compiler;
+
+use CoreShop\Bundle\PimcoreBundle\DependencyInjection\Compiler\RegisterRegistryTypePass;
+
+final class ExportProviderRegistryCompilerPass extends RegisterRegistryTypePass
+{
+    public function __construct()
+    {
+        parent::__construct(
+            'import_definition.registry.export_provider',
+            'import_definition.form.registry.export_provider',
+            'import_definition.export_providers',
+            'import_definition.export_provider'
+        );
+    }
+}
