+++ src/ImportDefinitionsBundle/ImportDefinitionsBundle.php
@@ -18,6 +18,7 @@ use CoreShop\Bundle\ResourceBundle\AbstractResourceBundle;
 use CoreShop\Bundle\ResourceBundle\ComposerPackageBundleInterface;
 use CoreShop\Bundle\ResourceBundle\CoreShopResourceBundle;
 use ImportDefinitionsBundle\DependencyInjection\Compiler\CleanerRegistryCompilerPass;
+use ImportDefinitionsBundle\DependencyInjection\Compiler\ExportProviderRegistryCompilerPass;
 use ImportDefinitionsBundle\DependencyInjection\Compiler\FetcherRegistryCompilerPass;
 use ImportDefinitionsBundle\DependencyInjection\Compiler\FilterRegistryCompilerPass;
 use ImportDefinitionsBundle\DependencyInjection\Compiler\GetterRegistryCompilerPass;
@@ -68,6 +69,7 @@ class ImportDefinitionsBundle extends AbstractResourceBundle implements PimcoreB
         $builder->addCompilerPass(new LoaderRegistryCompilerPass());
         $builder->addCompilerPass(new GetterRegistryCompilerPass());
         $builder->addCompilerPass(new FetcherRegistryCompilerPass());
+        $builder->addCompilerPass(new ExportProviderRegistryCompilerPass());
     }
 
     /**
