+++ src/ImportDefinitionsBundle/Interpreter/IteratorInterpreter.php
@@ -18,7 +18,7 @@ use CoreShop\Component\Registry\ServiceRegistryInterface;
 use ImportDefinitionsBundle\Model\DataSetAwareInterface;
 use ImportDefinitionsBundle\Model\DataSetAwareTrait;
 use ImportDefinitionsBundle\Model\DefinitionInterface;
-use ImportDefinitionsBundle\Model\MappingInterface;
+use ImportDefinitionsBundle\Model\Mapping;
 use Pimcore\Model\DataObject\Concrete;
 use Webmozart\Assert\Assert;
 
@@ -42,7 +42,7 @@ final class IteratorInterpreter implements InterpreterInterface, DataSetAwareInt
     /**
      * {@inheritdoc}
      */
-    public function interpret(Concrete $object, $value, MappingInterface $map, $data, DefinitionInterface $definition, $params, $configuration)
+    public function interpret(Concrete $object, $value, Mapping $map, $data, DefinitionInterface $definition, $params, $configuration)
     {
         Assert::isArray($value, 'IteratorInterpreter can only be used with array values');
