+++ src/ImportDefinitionsBundle/Interpreter/TypeCastingInterpreter.php
@@ -17,13 +17,12 @@ namespace ImportDefinitionsBundle\Interpreter;
 use ImportDefinitionsBundle\Model\DataSetAwareInterface;
 use ImportDefinitionsBundle\Model\DataSetAwareTrait;
 use ImportDefinitionsBundle\Model\DefinitionInterface;
-use ImportDefinitionsBundle\Model\MappingInterface;
+use ImportDefinitionsBundle\Model\Mapping;
 use Pimcore\Model\DataObject\Concrete;
 
 class TypeCastingInterpreter implements InterpreterInterface, DataSetAwareInterface
 {
     use DataSetAwareTrait;
-
 
     const TYPE_INT = 'int';
 
@@ -34,7 +33,7 @@ class TypeCastingInterpreter implements InterpreterInterface, DataSetAwareInterf
     /**
      * {@inheritdoc}
      */
-    public function interpret(Concrete $object, $value, MappingInterface $map, $data, DefinitionInterface $definition, $params, $configuration)
+    public function interpret(Concrete $object, $value, Mapping $map, $data, DefinitionInterface $definition, $params, $configuration)
     {
         $type = $configuration['toType'];
